[lib] use "0" as klass id for endianness event
Endianness event is sent before HawkTracer knows the endianness of
a profiled system, so the parser might not be able to recognize the ID
correctly (because it doesn't know anything about the system). In order
to parse such event without knowing the endianness, the klass id must
be a value which is the same for both Little and Big endian. One of such
values is zero, so we use it now as a klass id for endianness event.