[lib] enable listener's auto-management
Since this commit HawkTracer allows user to enable auto-managing the listener.
This means, that when the listener is created and registered to the timeline,
it doesn't have to be explicitly deleted in user's code.
In order to achieve that, user must provide a destroy callback, which is
called when the listener is no longer needed.