[lib] add option to force-compile HawkTracer library with C compilier
HawkTracer core library by default is compiled with C and C++ compiler
(depending on the source file), but some all the C++ files can be compiled
with C compiler as they have necessary ifdef's inside. This flag can be
particularly useful if taking a dependency to stdc++ is not wanted.